Stable

+++ app/models/calendar_event.rb
@@ -1,5 +1,5 @@
 #
-# Copyright (C) 2011 - 2014 Instructure, Inc.
+# Copyright (C) 2011 - present Instructure, Inc.
 #
 # This file is part of Canvas.
 #
@@ -20,12 +20,13 @@ require 'atom'
 require 'date'
 require 'icalendar'
 
-Icalendar::Event.ical_property  :x_alt_desc
+Icalendar::Event.optional_property  :x_alt_desc
 
 class CalendarEvent < ActiveRecord::Base
   include CopyAuthorizedLinks
   include TextHelper
   include HtmlTextHelper
+  include Plannable
 
   include MasterCourses::Restrictor
   restrict_columns :content, [:title, :description]
@@ -93,7 +94,7 @@ class CalendarEvent < ActiveRecord::Base
     return unless @child_event_data
     current_events = child_events.group_by{ |e| e[:context_code] }
     @child_event_data.each do |data|
-      if event = current_events.delete(data[:context_code]) and event = event[0]
+      if (event = current_events.delete(data[:context_code])&.first)
         event.updating_user = @updating_user
         event.update_attributes(:start_at => data[:start_at], :end_at => data[:end_at])
       else
@@ -118,8 +119,6 @@ class CalendarEvent < ActiveRecord::Base
   def effective_context
     effective_context_code && ActiveRecord::Base.find_by_asset_string(effective_context_code) || context
   end
-
-  scope :order_by_start_at, -> { order(:start_at) }
 
   scope :active, -> { where("calendar_events.workflow_state<>'deleted'") }
   scope :are_locked, -> { where(:workflow_state => 'locked') }
@@ -165,6 +164,10 @@ class CalendarEvent < ActiveRecord::Base
     SQL
   }
 
+  scope :not_hidden, -> {
+    where("NOT EXISTS (SELECT id FROM #{CalendarEvent.quoted_table_name} sub_events WHERE sub_events.parent_calendar_event_id=calendar_events.id)")
+  }
+
   scope :undated, -> { where(:start_at => nil, :end_at => nil) }
 
   scope :between, lambda { |start, ending| where(:start_at => start..ending) }
@@ -204,6 +207,29 @@ class CalendarEvent < ActiveRecord::Base
   end
   protected :default_values
 
+  def root_account
+    if context.respond_to?(:root_account)
+      context.root_account # course, section, group
+    else
+      case context
+      when User
+        context.account
+      when AppointmentGroup
+        context.context&.root_account
+      end
+    end
+  end
+
+  def child_event_participants_scope
+    self.shard.activate do
+      # user is set directly, or context is user
+      User.where("id IN
+        (#{child_events.where.not(:user_id => nil).select(:user_id).to_sql}
+        UNION
+         #{child_events.where(:user_id => nil, :context_type => "User").select(:context_id).to_sql})")
+    end
+  end
+
   def populate_appointment_group_defaults
     self.effective_context_code = context.appointment_group_contexts.map(&:context_code).join(",")
     if new_record?
@@ -242,18 +268,15 @@ class CalendarEvent < ActiveRecord::Base
   def populate_all_day_flag
     # If the all day flag has been changed to all day, set the times to 00:00
     if self.all_day_changed? && self.all_day?
-      self.start_at = self.end_at = zoned_start_at.beginning_of_day rescue nil
-
+      self.start_at = zoned_start_at.beginning_of_day rescue nil
+      self.end_at = zoned_end_at.beginning_of_day rescue nil
     elsif self.start_at_changed? || self.end_at_changed?
-      if self.start_at && self.start_at == self.end_at && zoned_start_at.strftime("%H:%M") == '00:00'
-        self.all_day = true
-      else
-        self.all_day = false
-      end
+      self.all_day = self.start_at && self.start_at == self.end_at && zoned_start_at.strftime("%H:%M") == '00:00'
     end
 
     if self.all_day && (!self.all_day_date || self.start_at_changed? || self.all_day_date_changed?)
-      self.start_at = self.end_at = zoned_start_at.beginning_of_day rescue nil
+      self.start_at = zoned_start_at.beginning_of_day rescue nil
+      self.end_at = zoned_end_at.beginning_of_day rescue nil
       self.all_day_date = (zoned_start_at.to_date rescue nil)
     end
   end
@@ -265,20 +288,25 @@ class CalendarEvent < ActiveRecord::Base
         ((ActiveSupport::TimeZone.new(self.time_zone_edited) rescue nil) || Time.zone))
   end
 
+  def zoned_end_at
+    self.end_at && ActiveSupport::TimeWithZone.new(self.end_at.utc,
+        ((ActiveSupport::TimeZone.new(self.time_zone_edited) rescue nil) || Time.zone))
+  end
+
   CASCADED_ATTRIBUTES = [
     :title,
     :description,
     :location_name,
     :location_address
-  ]
+  ].freeze
   LOCKED_ATTRIBUTES = CASCADED_ATTRIBUTES + [
     :start_at,
     :end_at
-  ]
+  ].freeze
 
   def sync_child_events
-    locked_changes = LOCKED_ATTRIBUTES.select { |attr| send("#{attr}_changed?") }
-    cascaded_changes = CASCADED_ATTRIBUTES.select { |attr| send("#{attr}_changed?") }
+    locked_changes = LOCKED_ATTRIBUTES.select { |attr| saved_change_to_attribute?(attr) }
+    cascaded_changes = CASCADED_ATTRIBUTES.select { |attr| saved_change_to_attribute?(attr) }
     child_events.are_locked.update_all Hash[locked_changes.map{ |attr| [attr, send(attr)] }] if locked_changes.present?
     child_events.are_unlocked.update_all Hash[cascaded_changes.map{ |attr| [attr, send(attr)] }] if cascaded_changes.present?
   end
@@ -287,13 +315,13 @@ class CalendarEvent < ActiveRecord::Base
   def sync_parent_event
     return unless parent_event
     return if appointment_group
-    return unless start_at_changed? || end_at_changed? || workflow_state_changed?
+    return unless saved_change_to_start_at? || saved_change_to_end_at? || saved_change_to_workflow_state?
     return if @skip_sync_parent_event
     parent_event.cache_child_event_ranges! unless workflow_state == 'deleted'
   end
 
   def cache_child_event_ranges!
-    events = child_events(true)
+    events = child_events.reload
 
     if events.present?
       CalendarEvent.where(:id => self).
@@ -368,7 +396,7 @@ class CalendarEvent < ActiveRecord::Base
       just_created &&
       context == appointment_group.participant_for(@updating_user)
     }
-    data { {:updating_user => @updating_user} }
+    data { {:updating_user_name => @updating_user.name} }
 
     dispatch :appointment_canceled_by_user
     to { appointment_group.instructors +
@@ -376,12 +404,12 @@ class CalendarEvent < ActiveRecord::Base
     whenever {
       appointment_group && parent_event &&
       deleted? &&
-      workflow_state_changed? &&
+      saved_change_to_workflow_state? &&
       @updating_user &&
       context == appointment_group.participant_for(@updating_user)
     }
     data { {
-      :updating_user => @updating_user,
+      :updating_user_name => @updating_user.name,
       :cancel_reason => @cancel_reason
     } }
 
@@ -391,17 +419,17 @@ class CalendarEvent < ActiveRecord::Base
       appointment_group && parent_event &&
       just_created
     }
-    data { {:updating_user => @updating_user} }
+    data { {:updating_user_name => @updating_user.name} }
 
     dispatch :appointment_deleted_for_user
     to { participants(include_observers: true) - [@updating_user] }
     whenever {
       appointment_group && parent_event &&
       deleted? &&
-      workflow_state_changed?
+      saved_change_to_workflow_state?
     }
     data { {
-      :updating_user => @updating_user,
+      :updating_user_name => @updating_user.name,
       :cancel_reason => @cancel_reason
     } }
   end
@@ -455,6 +483,7 @@ class CalendarEvent < ActiveRecord::Base
 
       if options[:cancel_existing]
         context.reservations_for(participant).lock.each do |reservation|
+          raise ReservationError, "cannot cancel past reservation" if reservation.end_at < Time.now.utc
           reservation.updating_user = user
           reservation.destroy
         end
@@ -530,10 +559,11 @@ class CalendarEvent < ActiveRecord::Base
     end
   end
 
-  def to_ics(in_own_calendar: true, preloaded_attachments: {}, user: nil)
+  def to_ics(in_own_calendar: true, preloaded_attachments: {}, user: nil, user_events: [])
     CalendarEvent::IcalEvent.new(self).to_ics(in_own_calendar:       in_own_calendar,
                                               preloaded_attachments: preloaded_attachments,
-                                              include_description:   true)
+                                              include_description:   true,
+                                              user_events: user_events)
   end
 
   def self.max_visible_calendars
@@ -589,61 +619,75 @@ class CalendarEvent < ActiveRecord::Base
     def location
     end
 
-    def to_ics(in_own_calendar:, preloaded_attachments: {}, include_description: false)
+    def to_ics(in_own_calendar:, preloaded_attachments: {}, include_description: false, user_events: [])
       cal = Icalendar::Calendar.new
       # to appease Outlook
-      cal.custom_property("METHOD","PUBLISH")
+      cal.append_custom_property("METHOD","PUBLISH")
 
       event = Icalendar::Event.new
-      event.klass = "PUBLIC"
+      event.ip_class = "PUBLIC"
 
       start_at = @event.is_a?(CalendarEvent) ? @event.start_at : @event.due_at
       end_at = @event.is_a?(CalendarEvent) ? @event.end_at : @event.due_at
 
-      if start_at
-        event.start = start_at.utc_datetime
-        event.start.icalendar_tzid = 'UTC'
-      end
+      event.dtstart = Icalendar::Values::DateTime.new(start_at.utc_datetime, 'tzid' => 'UTC') if start_at
+      event.dtend = Icalendar::Values::DateTime.new(end_at.utc_datetime, 'tzid' => 'UTC') if end_at
 
-      if end_at
-        event.end = end_at.utc_datetime
-        event.end.icalendar_tzid = 'UTC'
-      end
-
-      if @event.all_day
-        event.start = Date.new(@event.all_day_date.year, @event.all_day_date.month, @event.all_day_date.day)
-        event.start.ical_params = {"VALUE"=>["DATE"]}
-        event.end = event.start
-        event.end.ical_params = {"VALUE"=>["DATE"]}
+      if @event.all_day && @event.all_day_date
+        event.dtstart = Date.new(@event.all_day_date.year, @event.all_day_date.month, @event.all_day_date.day)
+        event.dtstart.ical_params = {"VALUE"=>["DATE"]}
+        event.dtend = event.dtstart
+        event.dtend.ical_params = {"VALUE"=>["DATE"]}
       end
 
       event.summary = @event.title
 
       if @event.description && include_description
         html = api_user_content(@event.description, @event.context, nil, preloaded_attachments)
-        event.description html_to_text(html)
-        event.x_alt_desc(html, { 'FMTTYPE' => 'text/html' })
+        event.description = html_to_text(html)
+        event.x_alt_desc = Icalendar::Values::Text.new(html, {'FMTTYPE' => 'text/html'})
       end
 
       if @event.is_a?(CalendarEvent)
-        loc_string = ""
-        loc_string << @event.location_name + ", " if @event.location_name.present?
-        loc_string << @event.location_address if @event.location_address.present?
+        loc_string = [@event.location_name, @event.location_address].reject { |e| e.blank? }.join(", ")
       else
         loc_string = nil
       end
 
+      if @event.context_type.eql?("AppointmentGroup")
+        # We should only enter this block if a user has made an appointment, so
+        # there is always at least one element in current_apts
+        current_appts = user_events.select { |appointment| @event.id == appointment[:parent_id]}
+        if current_appts.any?
+          if !event.description.nil?
+            event.description.concat("\n\n" + current_appts[0][:course_name] + "\n\n")
+          else
+            event.description = current_appts[0][:course_name] + "\n\n"
+          end
+
+          event.description.concat("Participants: ")
+          current_appts.each { |appt| event.description.concat("\n" + appt[:user]) }
+          comments = current_appts.map{ |appt| appt[:comments] }.join(",\n")
+          event.description.concat("\n\n" + comments)
+        end
+      end
+
       event.location = loc_string
-      event.dtstamp = @event.updated_at.utc_datetime if @event.updated_at
-      event.dtstamp.icalendar_tzid = 'UTC' if event.dtstamp
+      event.dtstamp = Icalendar::Values::DateTime.new(@event.updated_at.utc_datetime, 'tzid' => 'UTC') if @event.updated_at
 
       tag_name = @event.class.name.underscore
 
+      # Covers the case for when personal calendar event is created so that HostUrl finds the correct UR:
+      url_context = @event.context
+      if url_context.is_a? User
+        url_context = url_context.account
+      end
+
       # This will change when there are other things that have calendars...
       # can't call calendar_url or calendar_url_for here, have to do it manually
-      event.url           "http://#{HostUrl.context_host(@event.context)}/calendar?include_contexts=#{@event.context.asset_string}&month=#{start_at.try(:strftime, "%m")}&year=#{start_at.try(:strftime, "%Y")}##{tag_name}_#{@event.id}"
-      event.uid           "event-#{tag_name.gsub('_', '-')}-#{@event.id}"
-      event.sequence      0
+      event.url =         "https://#{HostUrl.context_host(url_context)}/calendar?include_contexts=#{@event.context.asset_string}&month=#{start_at.try(:strftime, "%m")}&year=#{start_at.try(:strftime, "%Y")}##{tag_name}_#{@event.id}"
+      event.uid =         "event-#{tag_name.gsub('_', '-')}-#{@event.id}"
+      event.sequence =    0
 
       if @event.respond_to?(:applied_overrides)
         @event.applied_overrides.try(:each) do |override|
