+++ app/models/importers/quiz_importer.rb
@@ -1,5 +1,5 @@
 #
-# Copyright (C) 2011 - 2013 Instructure, Inc.
+# Copyright (C) 2014 - present Instructure, Inc.
 #
 # This file is part of Canvas.
 #
@@ -166,27 +166,28 @@ module Importers
       # there might not be an import id if it's just a text-only type...
       item ||= Quizzes::Quiz.where(context_type: context.class.to_s, context_id: context, id: hash[:id]).first if hash[:id]
       item ||= Quizzes::Quiz.where(context_type: context.class.to_s, context_id: context, migration_id: hash[:migration_id]).first if hash[:migration_id]
-      if item && !allow_update
-        if item.deleted?
+      item ||= context.quizzes.temp_record
+      item.mark_as_importing!(migration)
+      if item
+        if !allow_update && item.deleted?
           item.workflow_state = (hash[:available] || !item.can_unpublish?) ? 'available' : 'unpublished'
           item.saved_by = :migration
           item.quiz_groups.destroy_all
-          item.quiz_questions.destroy_all
+          item.quiz_questions.preload(assessment_question: :assessment_question_bank).destroy_all
           item.save
         end
       end
-      item ||= context.quizzes.temp_record
-      item.mark_as_importing!(migration)
       new_record = item.new_record? || item.deleted?
 
-      hash[:due_at] ||= hash[:due_date]
+      hash[:due_at] ||= hash[:due_date] if hash.has_key?(:due_date)
       hash[:due_at] ||= hash[:grading][:due_date] if hash[:grading]
-      item.lock_at = Canvas::Migration::MigratorHelper.get_utc_time_from_timestamp(hash[:lock_at]) if hash[:lock_at]
-      item.unlock_at = Canvas::Migration::MigratorHelper.get_utc_time_from_timestamp(hash[:unlock_at]) if hash[:unlock_at]
-      item.due_at = Canvas::Migration::MigratorHelper.get_utc_time_from_timestamp(hash[:due_at]) if hash[:due_at]
-      item.show_correct_answers_at = Canvas::Migration::MigratorHelper.get_utc_time_from_timestamp(hash[:show_correct_answers_at]) if hash[:show_correct_answers_at]
-      item.hide_correct_answers_at = Canvas::Migration::MigratorHelper.get_utc_time_from_timestamp(hash[:hide_correct_answers_at]) if hash[:hide_correct_answers_at]
-      item.scoring_policy = hash[:which_attempt_to_keep] if hash[:which_attempt_to_keep]
+      master_migration = migration&.for_master_course_import? # propagate null dates only for blueprint syncs
+      item.lock_at = Canvas::Migration::MigratorHelper.get_utc_time_from_timestamp(hash[:lock_at]) if master_migration || hash[:lock_at]
+      item.unlock_at = Canvas::Migration::MigratorHelper.get_utc_time_from_timestamp(hash[:unlock_at]) if master_migration || hash[:unlock_at]
+      item.due_at = Canvas::Migration::MigratorHelper.get_utc_time_from_timestamp(hash[:due_at]) if master_migration || hash[:due_at]
+      item.show_correct_answers_at = Canvas::Migration::MigratorHelper.get_utc_time_from_timestamp(hash[:show_correct_answers_at]) if master_migration || hash[:show_correct_answers_at]
+      item.hide_correct_answers_at = Canvas::Migration::MigratorHelper.get_utc_time_from_timestamp(hash[:hide_correct_answers_at]) if master_migration || hash[:hide_correct_answers_at]
+      item.scoring_policy = hash[:which_attempt_to_keep] if master_migration || hash[:which_attempt_to_keep]
 
       missing_links = []
       item.description = migration.convert_html(hash[:description], :quiz, hash[:migration_id], :description)
@@ -217,42 +218,18 @@ module Importers
         only_visible_to_overrides
       ].each do |attr|
         attr = attr.to_sym
-        item.send("#{attr}=", hash[attr]) if hash.key?(attr)
+        if hash.key?(attr)
+          item.send("#{attr}=", hash[attr])
+        elsif master_migration
+          item.send("#{attr}=", nil)
+        end
       end
 
       item.saved_by = :migration
       item.save!
       build_assignment = false
 
-      skip_questions = migration.for_master_course_import? && item.edit_types_locked_for_overwrite_on_import.include?(:content)
-      if question_data && !skip_questions
-        question_data[:qq_ids] ||= {}
-        hash[:questions] ||= []
-
-        unless question_data[:qq_ids][item.migration_id]
-          question_data[:qq_ids][item.migration_id] = {}
-          existing_questions = item.quiz_questions.active.where("migration_id IS NOT NULL").pluck(:id, :migration_id)
-          existing_questions.each do |id, mig_id|
-            question_data[:qq_ids][item.migration_id][mig_id] = id
-          end
-        end
-
-        hash[:questions].each_with_index do |question, i|
-          case question[:question_type]
-          when "question_reference"
-            if aq = (question_data[:aq_data][question[:migration_id]] || question_data[:aq_data][question[:assessment_question_migration_id]])
-              Importers::QuizQuestionImporter.import_from_migration(aq, question, i + 1,
-                question_data[:qq_ids][item.migration_id], context, migration, item)
-            end
-          when "question_group"
-            Importers::QuizGroupImporter.import_from_migration(question, context, item, question_data, i + 1, migration)
-          when "text_only_question"
-            Importers::QuizQuestionImporter.import_from_migration(question, question, i + 1,
-              question_data[:qq_ids][item.migration_id], context, migration, item)
-          end
-        end
-      end
-      item.reload # reload to catch question additions
+      self.import_questions(item, hash, context, migration, question_data, new_record)
 
       if hash[:assignment]
         if hash[:assignment][:migration_id] && !hash[:assignment][:migration_id].start_with?(MasterCourses::MIGRATION_ID_PREFIX)
@@ -262,11 +239,6 @@ module Importers
         item.assignment ||= context.assignments.temp_record
 
         item.assignment = ::Importers::AssignmentImporter.import_from_migration(hash[:assignment], context, migration, item.assignment, item)
-
-        if new_record && !hash[:available] && item.can_unpublish?
-          item.workflow_state = 'unpublished'
-          item.assignment.workflow_state = 'unpublished'
-        end
       elsif !item.assignment && grading = hash[:grading]
         item.quiz_type = 'assignment'
         hash[:assignment_group_migration_id] ||= grading[:assignment_group_migration_id]
@@ -299,6 +271,9 @@ module Importers
         item.generate_quiz_data
         item.workflow_state = 'available'
         item.published_at = Time.now
+      elsif item.can_unpublish? && (new_record || master_migration)
+        item.workflow_state = 'unpublished'
+        item.assignment.workflow_state = 'unpublished' if item.assignment
       end
 
       if hash[:assignment_group_migration_id]
@@ -316,8 +291,10 @@ module Importers
         item.assignment.points_possible = item.points_possible
       end
 
+      item.root_entries(true) if !item.available? && !item.survey? # reload items so we get accurate points
+      item.notify_of_update = false
       item.save
-      item.assignment.save if item.assignment && item.assignment.changed?
+      item.assignment.save_without_broadcasting if item.assignment && item.assignment.changed?
 
       migration.add_imported_item(item)
       item.saved_by = nil
@@ -325,5 +302,50 @@ module Importers
       item
     end
 
+    def self.import_questions(item, hash, context, migration, question_data, new_record)
+      if migration.for_master_course_import? && !new_record
+        return if item.edit_types_locked_for_overwrite_on_import.include?(:content)
+
+        if hash[:questions]
+          # either the quiz hasn't been changed downstream or we've re-locked it - delete all the questions/question_groups we're not going to (re)import in
+          importing_qgroup_mig_ids = hash[:questions].select{|q| q[:question_type] == "question_group"}.map{|qg| qg[:migration_id]}
+          item.quiz_groups.where.not(:migration_id => importing_qgroup_mig_ids).destroy_all
+
+          importing_question_mig_ids = hash[:questions].map{|q| q[:questions] ?
+            q[:questions].map{|qq| qq[:quiz_question_migration_id] || qq[:migration_id]} :
+            q[:quiz_question_migration_id] || q[:migration_id]
+          }.flatten
+          item.quiz_questions.active.where.not(:migration_id => importing_question_mig_ids).update_all(:workflow_state => 'deleted')
+        end
+      end
+      return unless question_data
+
+      question_data[:qq_ids] ||= {}
+      hash[:questions] ||= []
+
+      unless question_data[:qq_ids][item.migration_id]
+        question_data[:qq_ids][item.migration_id] = {}
+        existing_questions = item.quiz_questions.active.where("migration_id IS NOT NULL").pluck(:id, :migration_id)
+        existing_questions.each do |id, mig_id|
+          question_data[:qq_ids][item.migration_id][mig_id] = id
+        end
+      end
+
+      hash[:questions].each_with_index do |question, i|
+        case question[:question_type]
+        when "question_reference"
+          if aq = (question_data[:aq_data][question[:migration_id]] || question_data[:aq_data][question[:assessment_question_migration_id]])
+            Importers::QuizQuestionImporter.import_from_migration(aq, question, i + 1,
+              question_data[:qq_ids][item.migration_id], context, migration, item)
+          end
+        when "question_group"
+          Importers::QuizGroupImporter.import_from_migration(question, context, item, question_data, i + 1, migration)
+        when "text_only_question"
+          Importers::QuizQuestionImporter.import_from_migration(question, question, i + 1,
+            question_data[:qq_ids][item.migration_id], context, migration, item)
+        end
+      end
+      item.reload # reload to catch question additions
+    end
   end
 end
